chore(user-emails): move user emails to core module
