+++ packages/server/modules/core/dbSchema.ts
@@ -600,3 +600,5 @@ export const GendoAIRenders = buildTableHelper('gendo_ai_renders', [
 ])
 
 export { knex }
+
+export const USER_EMAILS_TABLE_NAME = 'user_emails'
