--- packages/server/modules/user-emails/constants.ts
@@ -1,1 +1,0 @@
-export const USER_EMAILS_TABLE_NAME = 'user_emails'
